AP_NavEKF: extend fusion fault reporting coverage

+++ libraries/AP_NavEKF/AP_NavEKF.h
@@ -201,7 +201,7 @@ public:
      6 = badly conditioned synthetic sideslip fusion
      7 = filter is not initialised
     */
-    void  getFilterFaults(uint8_t &faults) const;
+    void  getFilterFaults(uint16_t &faults) const;
 
     /*
     return filter timeout status as a bitmasked integer
